Fixed bugs relating to probes.
* Simulator n_steps wasn't getting updated each step.

* Using unsigned long long ints as the key type to avoid overflow.

+++ mpi_sim/simulator.hpp
@@ -9,7 +9,7 @@
 
 using namespace std;
 
-typedef int key_type;
+typedef unsigned long long int key_type;
 
 //TODO: add a probing system
 class MpiSimulatorChunk{
